refactor(next-drupal)!: move non-DrupalClient code to src/deprecated
BREAKING CHANGE:
The JsonApiWithAuthOptions type has been renamed to JsonApiWithAuthOption. The
JsonApiWithLocaleOptions type is now considered deprecated for DrupalClient
usage and can be replaced with the JsonApiOptions type.
