+++ packages/next-drupal/src/deprecated.ts
@@ -1,0 +1,19 @@
+export * from "./deprecated/get-access-token"
+export * from "./deprecated/get-menu"
+export * from "./deprecated/get-paths"
+export * from "./deprecated/get-resource-collection"
+export * from "./deprecated/preview"
+export * from "./deprecated/get-resource-type"
+export * from "./deprecated/get-resource"
+export * from "./deprecated/get-search-index"
+export * from "./deprecated/get-view"
+export * from "./deprecated/translate-path"
+export {
+  deserialize,
+  buildUrl,
+  getJsonApiIndex,
+  getJsonApiPathForResourceType,
+  syncDrupalPreviewRoutes,
+} from "./deprecated/utils"
+
+export type * from "./types/deprecated"
