+++ packages/next-drupal/src/deprecated/get-resource-type.ts
@@ -1,6 +1,6 @@
 import { translatePathFromContext } from "./translate-path"
 import type { GetStaticPropsContext } from "next"
-import type { AccessToken } from "./types"
+import type { AccessToken } from "../types"
 
 export async function getResourceTypeFromContext(
   context: GetStaticPropsContext,
