+++ packages/next-drupal/src/deprecated/get-resource.ts
@@ -7,12 +7,8 @@ import {
   getPathFromContext,
 } from "./utils"
 import type { GetStaticPropsContext } from "next"
-import type {
-  AccessToken,
-  JsonApiParams,
-  JsonApiResource,
-  JsonApiWithLocaleOptions,
-} from "./types"
+import type { AccessToken, JsonApiParams, JsonApiResource } from "../types"
+import type { JsonApiWithLocaleOptions } from "../types/deprecated"
 
 export async function getResourceFromContext<T extends JsonApiResource>(
   type: string,
