+++ packages/next-drupal/src/deprecated/get-search-index.ts
@@ -1,10 +1,7 @@
 import { buildHeaders, buildUrl, deserialize } from "./utils"
 import type { GetStaticPropsContext } from "next"
-import type {
-  AccessToken,
-  JsonApiResource,
-  JsonApiWithLocaleOptions,
-} from "./types"
+import type { AccessToken, JsonApiResource } from "../types"
+import type { JsonApiWithLocaleOptions } from "../types/deprecated"
 
 export async function getSearchIndex<T = JsonApiResource[]>(
   name: string,
