+++ packages/next-drupal/src/deprecated/use-menu.ts
@@ -1,6 +1,6 @@
 import { useRouter } from "next/router"
 import { useEffect, useState } from "react"
-import { getMenu } from "../get-menu"
+import { getMenu } from "./get-menu"
 import type { DrupalMenuLinkContent } from "../types"
 
 export function useMenu<T extends DrupalMenuLinkContent>(
