+++ packages/next-drupal/src/index.ts
@@ -1,21 +1,6 @@
-export * from "./get-access-token"
-export * from "./get-menu"
-export * from "./get-paths"
-export * from "./get-resource-collection"
-export * from "./preview"
-export * from "./get-resource-type"
-export * from "./get-resource"
-export * from "./get-search-index"
-export * from "./get-view"
-export * from "./types"
-export * from "./translate-path"
-export {
-  deserialize,
-  buildUrl,
-  getJsonApiIndex,
-  getJsonApiPathForResourceType,
-  syncDrupalPreviewRoutes,
-} from "./utils"
-
 export * from "./client"
 export * from "./jsonapi-errors"
+
+export type * from "./types"
+
+export * from "./deprecated"
