+++ packages/next-drupal/src/jsonapi-errors.ts
@@ -1,4 +1,17 @@
-import type { JsonApiError } from "./types"
+// https://jsonapi.org/format/#error-objects
+export interface JsonApiError {
+  id?: string
+  status?: string
+  code?: string
+  title?: string
+  detail?: string
+  links?: JsonApiLinks
+}
+
+// https://jsonapi.org/format/#document-links
+export interface JsonApiLinks {
+  [key: string]: string | Record<string, string>
+}
 
 export class JsonApiErrors extends Error {
   errors: JsonApiError[] | string
