+++ packages/next-drupal/src/types/deprecated.ts
@@ -1,0 +1,3 @@
+import type { JsonApiOptions } from "./index"
+
+export type JsonApiWithLocaleOptions = Omit<JsonApiOptions, "withAuth">
