+++ packages/next-drupal/src/types/index.ts
@@ -1,0 +1,4 @@
+export type * from "./client"
+export type * from "./drupal"
+export type * from "./options"
+export type * from "./resource"
